eliminate 64-to-32 implicit conversions

+++ src/tables/otl/gpos-mark-to-single.c
@@ -85,7 +85,7 @@ json_value *caryll_gpos_mark_to_single_to_json(otl_subtable *st) {
 		sds markClassName =
 		    sdscatfmt(sdsempty(), "ac_%i", subtable->markArray->records[j].markClass);
 		json_object_push(_mark, "class",
-		                 json_string_new_length(sdslen(markClassName), markClassName));
+		                 json_string_new_length((uint32_t)sdslen(markClassName), markClassName));
 		sdsfree(markClassName);
 		json_object_push(_mark, "x", json_integer_new(subtable->markArray->records[j].anchor.x));
 		json_object_push(_mark, "y", json_integer_new(subtable->markArray->records[j].anchor.y));
@@ -99,7 +99,8 @@ json_value *caryll_gpos_mark_to_single_to_json(otl_subtable *st) {
 				json_object_push(_anchor, "x", json_integer_new(subtable->baseArray[j][k].x));
 				json_object_push(_anchor, "y", json_integer_new(subtable->baseArray[j][k].y));
 				sds markClassName = sdscatfmt(sdsempty(), "ac_%i", k);
-				json_object_push_length(_base, sdslen(markClassName), markClassName, _anchor);
+				json_object_push_length(_base, (uint32_t)sdslen(markClassName), markClassName,
+				                        _anchor);
 				sdsfree(markClassName);
 			}
 		}
